+++ apps/server/src/sessions/dto/create-session.dto.ts
@@ -1,5 +1,6 @@
+import { SessionRoleType } from "@prisma/client"
 import { Transform, TransformFnParams } from "class-transformer"
-import { IsArray, IsBoolean, IsNotEmpty, IsNumber, IsOptional, IsString, Matches } from "class-validator"
+import { IsArray, IsBoolean, IsEnum, IsNotEmpty, IsNumber, IsOptional, IsString, Matches } from "class-validator"
 import { IsRangeTime } from "src/validators/is-range-time.validator"
 import { IsTime } from "src/validators/is-time.validator"
 
@@ -30,6 +31,11 @@ export class CreateSessionDto {
     @IsString({ each: true })
     @IsOptional()
     group: string[]
+
+    @IsString({ message: "Session Role Type Harus String"})
+    @IsEnum(SessionRoleType, { message: "Session Role Type Salah, Hanya Ada 2 Yaitu Siswa dan Pegawai" })
+    @IsOptional()
+    session_role_type: SessionRoleType
 }